backend/requirements.txt
django
djangorestframework
django-cors-headers
pytest
shortuuid
isoweek
pendulum